Strong params for tickets are now in a rails concern
