+++ app/controllers/tickets_controller.rb
@@ -16,6 +16,7 @@
 
 class TicketsController < ApplicationController
   include HtmlTextHelper
+  include TicketsStrongParams
   include ActionView::Helpers::SanitizeHelper # dependency of HtmlTextHelper
 
   before_filter :authenticate_user!, except: [:create, :new]
@@ -178,33 +179,4 @@ class TicketsController < ApplicationController
       format.js { render }
     end
   end
-
-  protected
-    def ticket_params
-      if !current_user.nil? && current_user.agent?
-        params.require(:ticket).permit(
-            :from,
-            :to_email_address_id,
-            :content,
-            :subject,
-            :status,
-            :assignee_id,
-            :priority,
-            :message_id,
-            :content_type,
-            attachments_attributes: [
-              :file
-            ])
-      else
-        params.require(:ticket).permit(
-            :from,
-            :content,
-            :subject,
-            :priority,
-            :content_type,
-            attachments_attributes: [
-              :file
-            ])
-      end
-    end
 end
